chore: mv `assigned_relaxed_plonk_instance` to file
